* *.*: Deprecating exult_types.h in favor of common_types.h. Fixing several   cases where Exult used longs instead of pointer-sized integers to convert   pointers to integers (or the reverse); this fails on 64-bit Windows.

+++ gameclk.h
@@ -102,7 +102,7 @@ public:
 	void set_overcast(bool onoff);  // Start/end cloud cover.
 	void set_fog(bool onoff);   // Start/end cloud cover.
 	void increment(int num_minutes);// Increment clock.
-	virtual void handle_event(unsigned long curtime, long udata);
+	virtual void handle_event(unsigned long curtime, uintptr udata);
 	void fake_next_period();    // For debugging.
 	int get_time_rate() {
 		return time_rate;
